# api/openapi.yaml
openapi: 3.1.1
info:
  title: CajeerEngine REST API
  version: 1.1.1
  summary: Рабочий REST API CajeerEngine 1.1.1 stable.
  description: |
    REST API между Backend Core, Admin UI, SDK и headless-frontends.
    В 1.1.1 добавлен полноценный Admin UI hardening: production static SPA, dashboard, CRUD-экраны CMS, media, users/RBAC, extensions, updates, search и import/export.
...
servers:
  - url: /api/v1
...
paths:
  /admin/bootstrap:
    get:
      tags: [Admin]
      operationId: get_AdminBootstrap
      summary: GET /admin/bootstrap
      responses:
        '200':
          description: OK
        '401':
          description: Требуется авторизация
        '403':
          description: Недостаточно прав
        '422':
          description: Ошибка валидации
  /admin/dashboard:
    get:
      tags: [Admin]
      operationId: get_AdminDashboard
      summary: GET /admin/dashboard
      responses:
        '200':
          description: OK
        '401':
          description: Требуется авторизация
        '403':
          description: Недостаточно прав
        '422':
          description: Ошибка валидации
  /api-tokens:
    get:
      tags: [Security]
      operationId: get_ApiTokensIndex
      summary: GET /api-tokens
      responses:
        '200':
          description: OK
        '401':
          description: Требуется авторизация
        '403':
          description: Недостаточно прав
        '422':
          description: Ошибка валидации
    post:
      tags: [Security]
      operationId: post_ApiTokensStore
      summary: POST /api-tokens
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '201':
          description: OK
        '401':
          description: Требуется авторизация
        '403':
          description: Недостаточно прав
        '422':
          description: Ошибка валидации
...
  /auth/login:
    post:
      tags: [Auth]
      operationId: post_AuthLogin
      summary: POST /auth/login
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '201':
          description: OK
        '401':
          description: Требуется авторизация
        '403':
          description: Недостаточно прав
        '422':
          description: Ошибка валидации
  /auth/logout:
    post:
      tags: [Auth]
      operationId: post_AuthLogout
      summary: POST /auth/logout
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '201':
          description: OK
        '401':
          description: Требуется авторизация
        '403':
          description: Недостаточно прав
        '422':
          description: Ошибка валидации
  /auth/me:
    get:
      tags: [Auth]
      operationId: get_AuthMe
      summary: GET /auth/me
      responses:
        '200':
          description: OK
        '401':
          description: Требуется авторизация
        '403':
          description: Недостаточно прав
        '422':
          description: Ошибка валидации
...
  /database:
    get:
      tags: [Database]
      operationId: get_DatabaseShow
      summary: GET /database
      responses:
        '200':
          description: OK
        '401':
          description: Требуется авторизация
        '403':
          description: Недостаточно прав
        '422':
          description: Ошибка валидации
...
  /extensions:
    get:
      tags: [Extensions]
      operationId: get_ExtensionsIndex
      summary: GET /extensions
      responses:
        '200':
          description: OK
        '401':
          description: Требуется авторизация
        '403':
          description: Недостаточно прав
        '422':
          description: Ошибка валидации
...
  /extensions/events:
    get:
      tags: [Extensions]
      operationId: get_ExtensionsEvents
      summary: GET /extensions/events
      responses:
        '200':
          description: OK
        '401':
          description: Требуется авторизация
        '403':
          description: Недостаточно прав
        '422':
          description: Ошибка валидации
...
  /media:
    get:
      tags: [Media]
      operationId: get_MediaIndex
      summary: GET /media
      responses:
        '200':
          description: OK
        '401':
          description: Требуется авторизация
        '403':
          description: Недостаточно прав
        '422':
          description: Ошибка валидации
    post:
      tags: [Media]
      operationId: post_MediaStore
      summary: POST /media
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '201':
          description: OK
        '401':
          description: Требуется авторизация
        '403':
          description: Недостаточно прав
        '422':
          description: Ошибка валидации
...
  /queue:
    get:
      tags: [Queue]
      operationId: get_QueueDiagnostics
      summary: GET /queue
      responses:
        '200':
          description: OK
        '401':
          description: Требуется авторизация
        '403':
          description: Недостаточно прав
        '422':
          description: Ошибка валидации
...
  /roles:
    get:
      tags: [Users]
      operationId: get_RolesIndex
      summary: GET /roles
      responses:
        '200':
          description: OK
        '401':
          description: Требуется авторизация
        '403':
          description: Недостаточно прав
        '422':
          description: Ошибка валидации
    post:
      tags: [Users]
      operationId: post_RolesStore
      summary: POST /roles
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '201':
          description: OK
        '401':
          description: Требуется авторизация
        '403':
          description: Недостаточно прав
        '422':
          description: Ошибка валидации
...
  /runtime:
    get:
      tags: [System]
      operationId: get_Runtime
      summary: GET /runtime
      responses:
        '200':
          description: OK
        '401':
          description: Требуется авторизация
        '403':
          description: Недостаточно прав
        '422':
          description: Ошибка валидации
...
  /search:
    get:
      tags: [Search]
      operationId: get_SearchIndex
      summary: GET /search
      responses:
        '200':
          description: OK
        '401':
          description: Требуется авторизация
        '403':
          description: Недостаточно прав
        '422':
          description: Ошибка валидации
...
  /system:
    get:
      tags: [System]
      operationId: get_System
      summary: GET /system
      responses:
        '200':
          description: OK
        '401':
          description: Требуется авторизация
        '403':
          description: Недостаточно прав
        '422':
          description: Ошибка валидации
  /updates:
    get:
      tags: [Updates]
      operationId: get_UpdatesDiagnostics
      summary: GET /updates
      responses:
        '200':
          description: OK
        '401':
          description: Требуется авторизация
        '403':
          description: Недостаточно прав
        '422':
          description: Ошибка валидации
...
  /users:
    get:
      tags: [Users]
      operationId: get_UsersIndex
      summary: GET /users
      responses:
        '200':
          description: OK
        '401':
          description: Требуется авторизация
        '403':
          description: Недостаточно прав
        '422':
          description: Ошибка валидации
    post:
      tags: [Users]
      operationId: post_UsersStore
      summary: POST /users
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '201':
          description: OK
        '401':
          description: Требуется авторизация
        '403':
          description: Недостаточно прав
        '422':
          description: Ошибка валидации
...
  /webhooks:
    get:
      tags: [Webhooks]
      operationId: get_WebhooksIndex
      summary: GET /webhooks
      responses:
        '200':
          description: OK
        '401':
          description: Требуется авторизация
        '403':
          description: Недостаточно прав
        '422':
          description: Ошибка валидации
    post:
      tags: [Webhooks]
      operationId: post_WebhooksStore
      summary: POST /webhooks
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '201':
          description: OK
        '401':
          description: Требуется авторизация
        '403':
          description: Недостаточно прав
        '422':
          description: Ошибка валидации
...
  /cms:
    get:
      tags: [CMS]
      operationId: get_CmsOverview
      summary: GET /cms
      responses:
        '200': { description: OK }
        '401': { description: Требуется авторизация }
        '403': { description: Недостаточно прав }
...
  /extensions/runtime:
    get:
      tags: [Extensions]
      operationId: get_ExtensionRuntime
      summary: Extension runtime diagnostics and optional boot check
      parameters:
        - name: boot
          in: query
          required: false
          schema: { type: boolean }
      responses:
        '200': { description: OK }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    apiToken:
      type: apiKey
      in: header
      name: X-API-Token
...
security:
  - bearerAuth: []
  - apiToken: []